Add confirmation at setup

+++ app/clean.py
@@ -1,10 +1,15 @@
 #!/usr/bin/env python3
 
+import sys
 from shutil import rmtree
 from modules.logging import log_dir
 from modules.database import connect
 from modules.crawling import wdm_dir
-from modules.utilities import mkdirp
+from modules.utilities import mkdirp, confirm_cui
+
+# 誤実行防止
+if not confirm_cui("クリーンシーケンスを実行しますか？"):
+	sys.exit(1)
 
 # WDMのキャッシュをクリア
 rmtree(wdm_dir, ignore_errors=True)
